fixed bug in block lib select hander; added comments to factory.js to group the functions
