+++ demos/blockfactory/index.html
@@ -18,10 +18,17 @@
   <script src="block_exporter_view.js"></script>
   <script src="block_exporter_controller.js"></script>
   <script src="blocks.js"></script>
-  <script src="block_factory_expansion.js"></script>
+  <script src="app_controller.js"></script>
   <link rel="stylesheet" href="factory.css">
   <link rel="stylesheet" href="../prettify.css">
   <script src="../prettify.js"></script>
+  <script>
+    var init = function () {
+      var blocklyDevTools = new AppController();
+      blocklyDevTools.init();
+    };
+    window.addEventListener('load', init);
+</script>
 </head>
 <body>
   <h1><a href="https://developers.google.com/blockly/">Blockly</a> &gt;
@@ -69,6 +76,7 @@
         <button id="clearSelectedButton"> Clear Selected Blocks </button>
     </div>
   </div>
+
   <table id="blockFactoryContent">
     <tr>
       <td width="50%" height="5%">
@@ -77,7 +85,7 @@
             <td id="blockLibraryContainer">
             <span>
               <h3>Block Library:</h3>
-              <select id="blockLibraryDropdown" onChange="BlockFactoryExpansion.onSelectedBlockChanged(this);">
+              <select id="blockLibraryDropdown">
               </select>
             </span>
             </td>
